// src/app/core/interfaces/bioimpedancia.ts
export interface Bioimpedancia {
    email: string
    dataBio: Date
    peso: number
    imc: number
    gorduraCorporal: number
    aguaCorporal: number
    massaEsqueletica: number
    tmb: number
    massaLivreGordura: number
    gorduraSubcutanea: number
    gorduraVisceral: number
    massaMuscular: number
    massaOssea: number
    proteina: number
    idadeMetabolica: number
}